requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.1
python-dotenv>=1.0.0
llama-index>=0.9.0
llama-index-embeddings-huggingface>=0.1.0
google-generativeai>=0.3.0
PyPDF2>=3.0.0
prometheus-client>=0.17.1 